# docs/swagger.yaml
basePath: /
definitions:
  github_com_cdcdx_high-concurrency-framework_internal_model.ApiResponse:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
      trace_id:
        type: string
    type: object
...
  github_com_cdcdx_high-concurrency-framework_internal_model.Order:
    properties:
      amount:
        type: number
      created_at:
        type: string
      id:
        type: integer
      order_no:
        type: string
      status:
        type: string
      updated_at:
        type: string
      user_id:
        minimum: 1
        type: integer
    required:
    - user_id
    type: object
...
  github_com_cdcdx_high-concurrency-framework_internal_model.UserProfile:
    properties:
      avatar:
        type: string
      bio:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      nickname:
        type: string
      phone:
        type: string
      tags:
        items:
          type: string
        type: array
      updated_at:
        type: string
      user_id:
        type: integer
    required:
    - user_id
    type: object
host: localhost:8080
info:
  contact:
    name: cdcdx
  description: 高并发业务处理框架，支持多级缓存、读写分离、异步消息、全文搜索等能力
  title: 高并发业务框架 API
  version: 1.0.0
paths:
...
  /api/v1/orders:
    post:
      consumes:
      - application/json
      description: 双通道分流：同步(限流内) | 异步(超限→Kafka)，适用于高并发写场景
      parameters:
      - description: 订单信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_cdcdx_high-concurrency-framework_internal_model.Order'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_cdcdx_high-concurrency-framework_internal_model.ApiResponse'
            - properties:
                data:
                  properties:
                    channel:
                      type: string
                    order_no:
                      type: string
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_cdcdx_high-concurrency-framework_internal_model.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_cdcdx_high-concurrency-framework_internal_model.ApiResponse'
      security:
      - Bearer: []
      summary: 创建订单 (异步)
      tags:
      - 订单
  /api/v1/orders/{orderNo}:
    get:
      description: 通过订单号查询，走L1→L2→MySQL三级缓存
      parameters:
      - description: 订单号
        in: path
        name: orderNo
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_cdcdx_high-concurrency-framework_internal_model.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_cdcdx_high-concurrency-framework_internal_model.Order'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_cdcdx_high-concurrency-framework_internal_model.ApiResponse'
      security:
      - Bearer: []
      summary: 查询订单
      tags:
      - 订单
  /api/v1/orders/search:
    get:
      description: 通过Elasticsearch全文搜索订单
      parameters:
      - description: 搜索关键词
        in: query
        name: q
        required: true
        type: string
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页条数
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_cdcdx_high-concurrency-framework_internal_model.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_cdcdx_high-concurrency-framework_internal_model.ApiResponse'
      security:
      - Bearer: []
      summary: 搜索订单
      tags:
      - 订单
  /api/v1/orders/sync:
    post:
      consumes:
      - application/json
      description: 强制同步写入MySQL，不走Kafka，立即可读
      parameters:
      - description: 订单信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_cdcdx_high-concurrency-framework_internal_model.Order'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_cdcdx_high-concurrency-framework_internal_model.ApiResponse'
            - properties:
                data:
                  properties:
                    channel:
                      type: string
                    order_no:
                      type: string
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_cdcdx_high-concurrency-framework_internal_model.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_cdcdx_high-concurrency-framework_internal_model.ApiResponse'
      security:
      - Bearer: []
      summary: 创建订单 (同步)
      tags:
      - 订单
  /api/v1/users/{userID}/profile:
    get:
      description: '走多级缓存: L1→L2→L3(MongoDB)'
      parameters:
      - description: 用户ID
        in: path
        name: userID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_cdcdx_high-concurrency-framework_internal_model.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_cdcdx_high-concurrency-framework_internal_model.UserProfile'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_cdcdx_high-concurrency-framework_internal_model.ApiResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_cdcdx_high-concurrency-framework_internal_model.ApiResponse'
      security:
      - Bearer: []
      summary: 查询用户资料
      tags:
      - 用户
  /api/v1/users/profile:
    post:
      consumes:
      - application/json
      description: MongoDB upsert 写入，同时更新缓存
      parameters:
      - description: 用户资料
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_cdcdx_high-concurrency-framework_internal_model.UserProfile'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_cdcdx_high-concurrency-framework_internal_model.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_cdcdx_high-concurrency-framework_internal_model.UserProfile'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_cdcdx_high-concurrency-framework_internal_model.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_cdcdx_high-concurrency-framework_internal_model.ApiResponse'
      security:
      - Bearer: []
      summary: 创建/更新用户资料
      tags:
      - 用户
  /api/v1/users/search:
    get:
      description: 通过Elasticsearch全文搜索用户
      parameters:
      - description: 搜索关键词
        in: query
        name: q
        required: true
        type: string
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页条数
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_cdcdx_high-concurrency-framework_internal_model.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_cdcdx_high-concurrency-framework_internal_model.ApiResponse'
      security:
      - Bearer: []
      summary: 搜索用户
      tags:
      - 用户
securityDefinitions:
  Bearer:
    description: 输入JWT令牌 (可带或不带 "Bearer " 前缀), 通过 POST /api/v1/auth/login 或 /api/v1/auth/register
      获取
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"